refactor: change throttle pin value return types to unsigned for consistency and improved accuracy

+++ src/Power/Power.cpp
@@ -11,8 +11,8 @@ unsigned int Power::getPwm() {
     unsigned int throttleRaw = throttle.getThrottleRaw();
     unsigned int powerLimit = getPower(); // 0-100
 
-    unsigned int throttleMin = (unsigned int)throttle.getThrottlePinMin();
-    unsigned int throttleMax = (unsigned int)throttle.getThrottlePinMax();
+    unsigned int throttleMin = throttle.getThrottlePinMin();
+    unsigned int throttleMax = throttle.getThrottlePinMax();
     unsigned int allowedMax = throttleMin + ((throttleMax - throttleMin) * powerLimit) / 100;
     unsigned int effectiveRaw = constrain(throttleRaw, throttleMin, allowedMax);
